pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Thingo"
authors = [
    {name = "Nigel Williams",email = "nigel.williams@marketplacer.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3 (>=1.40.21,<2.0.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "strands-agents-tools (>=0.2.5,<0.3.0)",
    "strands-agents-builder (>=0.1.8,<0.2.0)",
    "aws-opentelemetry-distro (>=0.12.0,<0.13.0)",
    "bedrock-agentcore (>=0.1.2,<0.2.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pylint = "^3.3.8"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
